+++ src/Dynamics365Configuration/Install-Dynamics365ReportingExtensions.psm1
@@ -125,7 +125,7 @@
         }
         if([String]::IsNullOrEmpty($logFilePath) -eq $True) {
             $timeStamp = ( Get-Date -Format u ).Replace(" ","-").Replace(":","-");
-            $logFilePath = "$env:Temp\DynamicsReportingExtensionsInstallationLog_$timeStamp.txt";
+            $logFilePath = "$env:APPDATA\Microsoft\MSCRM\Logs\DynamicsReportingExtensionsInstallationLog_$timeStamp.txt";
         }
         Invoke-Command -ScriptBlock $localInstallationScriptBlock `
             -ArgumentList $setupFilePath, $stringWriter.ToString(), $logFilePath, $logFilePullIntervalInSeconds, $logFilePullToOutput;