+++ src/Dynamics365Configuration/Install-Dynamics365Server.psm1
@@ -539,7 +539,7 @@
                     }
                     if([String]::IsNullOrEmpty($logFilePath) -eq $True) {
                         $timeStamp = ( Get-Date -Format u ).Replace(" ","-").Replace(":","-");
-                        $logFilePath = "$env:Temp\CRMInstallationLog_$timeStamp.txt";
+                        $logFilePath = "$env:APPDATA\Microsoft\MSCRM\Logs\CRMInstallationLog_$timeStamp.txt";
                     }
                     Invoke-Command -ScriptBlock $localInstallationScriptBlock `
                         -ArgumentList $setupFilePath, $stringWriter.ToString(), $logFilePath, $logFilePullIntervalInSeconds, $logFilePullToOutput;