+++ src/Dynamics365Configuration/Install-Dynamics365Update.psm1
@@ -39,7 +39,7 @@
     }
     if([String]::IsNullOrEmpty($logFilePath) -eq $True) {
         $timeStamp = ( Get-Date -Format u ).Replace(" ","-").Replace(":","-");
-        $logFilePath = "$env:Temp\DynamicsUpdateInstallationLog_$timeStamp.txt";
+        $logFilePath = "$env:APPDATA\Microsoft\MSCRM\Logs\DynamicsUpdateInstallationLog_$timeStamp.txt";
     }
 
     Write-Output "$(Get-Date) Starting $setupFilePath";